backend/requirements.txt
Flask==2.0.1
numpy==1.24.3
pandas==2.0.3
psycopg2-binary==2.9.1
python-dotenv==0.19.0